docker

+++ password/src/main/java/esprit/wd/password/service/PasswordResetService.java
@@ -1,14 +1,13 @@
 package esprit.wd.password.service;
 
 
-import esprit.wd.password.exception.UserNotFoundException;
+import esprit.wd.password.dto.UserRequest;
 import esprit.wd.password.model.PasswordResetToken;
-import esprit.wd.password.model.User;
 import esprit.wd.password.repository.PasswordResetTokenRepository;
+import esprit.wd.password.request.ForgotPassword;
 import esprit.wd.password.user_client.UserServiceClient;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.stereotype.Service;
 
 import java.util.Date;
@@ -25,14 +24,13 @@ public class PasswordResetService {
     private final UserServiceClient userServiceClient;
 
     public String forgotPassword(String email) {
-        var user = userServiceClient.getUserByMail(email);
-        System.out.println("user: " + user);
+        var userId = userServiceClient.getUserByMail(email);
 
         String token = UUID.randomUUID().toString();
-        PasswordResetToken passwordResetToken = new PasswordResetToken(token, user.getUserId());
+        PasswordResetToken passwordResetToken = new PasswordResetToken(token, userId);
         tokenRepository.save(passwordResetToken);
 
-        emailService.sendResetPasswordEmail(passwordResetToken);
+        emailService.sendResetPasswordEmail(passwordResetToken, email);
 
         return "Password reset email sent!";
     }
@@ -42,16 +40,22 @@ public class PasswordResetService {
         return passwordResetToken.isPresent() && passwordResetToken.get().getExpiryDate().after(new Date());
     }
 
-    public String resetPassword(String token, String newPassword) {
+    public String resetPassword(String token, ForgotPassword request) {
 
-        PasswordResetToken resetToken = tokenRepository.findByToken(token)
-                .orElseThrow(() -> new RuntimeException("Invalid token"));
+        if (request.confirmNewPassword().equals(request.newPassword())) {
+            PasswordResetToken resetToken = tokenRepository.findByToken(token)
+                    .orElseThrow(() -> new RuntimeException("Invalid token"));
 
-        User user = userServiceClient.getUserByUserId(resetToken.getUserId());
-        user.setPassword(new BCryptPasswordEncoder().encode(newPassword));
-        userServiceClient.updateUserPassword(user);
-        tokenRepository.delete(resetToken);
 
-        return "Password has been reset successfully";
+            var user = UserRequest.builder()
+                    .userId(resetToken.getUserId())
+                    .password(request.newPassword().trim())
+                    .build();
+            userServiceClient.updateUserPassword(user);
+            tokenRepository.delete(resetToken);
+
+            return "Password has been reset successfully";
+        }
+        throw new RuntimeException("Invalid or expired token or password");
     }
 }
